Separated SkinInstance class into its own file skin-instance.js (#2394)
