+++ src/index.js
@@ -90,7 +90,8 @@ export { MorphTarget } from './scene/morph-target.js';
 export { ParticleEmitter } from './scene/particle-system/particle-emitter.js';
 export { Picker } from './scene/pick.js';
 export { Scene } from './scene/scene.js';
-export { Skin, SkinInstance } from './scene/skin.js';
+export { Skin } from './scene/skin.js';
+export { SkinInstance } from './scene/skin-instance.js';
 export { Sprite } from './scene/sprite.js';
 export { StandardMaterial } from './scene/materials/standard-material.js';
 export { StencilParameters } from './scene/stencil-parameters.js';
