+++ src/scene/batching.js
@@ -17,7 +17,7 @@ import { LAYERID_WORLD, SPRITE_RENDERMODE_SIMPLE } from './constants.js';
 import { Mesh } from './mesh.js';
 import { MeshInstance } from './mesh-instance.js';
 import { Model } from './model.js';
-import { SkinInstance } from './skin.js';
+import { SkinInstance } from './skin-instance.js';
 
 /**
  * @class
